// src/types/category.ts
export enum FoodCategory {
  Vorspeise = "Vorspeise",
  Hauptgericht = "Hauptgericht",
  Nachspeise = "Nachspeise",
  Beilage = "Beilage",
  Salat = "Salat",
  Suesses = "Süßes",
}

export enum DrinkCategory {
  Bier = "Bier",
  Wein = "Wein",
  Softdrinks = "Softdrinks",
  Heissgetraenke = "Heißgetränke",
  Spirituosen = "Spirituosen",
}